+++ monitor2.h
@@ -43,8 +43,8 @@ typedef struct {
     unsigned int TCPMaxWindowSize; //only TCP
     unsigned int RxRing;
     unsigned int TxRing;
-    u_int8_t nqaCnt1;
-    u_int8_t nqaCnt5;
+    unsigned int nqaCnt1;
+    unsigned int nqaCnt5;
 } networkData;
 
 typedef struct {
